Flytt arbeidserfaringState til egen fil
