+++ src/veileder/AppState.ts
@@ -1,6 +1,7 @@
 import { KandidatlisteState } from './kandidatlister/reducer/kandidatlisteReducer';
 import { PermitteringState } from './sok/permittering/permitteringReducer';
 import { OppstartstidspunktState } from './sok/oppstardstidspunkt/oppstartstidspunktReducer';
+import { ArbeidserfaringState, TypeaheadState } from './sok/arbeidserfaring/arbeidserfaringState';
 
 type AppState = {
     kandidatlister: KandidatlisteState;
@@ -10,28 +11,5 @@ type AppState = {
     arbeidserfaring: ArbeidserfaringState;
     typeahead: TypeaheadState;
 };
-
-interface ArbeidserfaringState {
-    arbeidserfaringPanelOpen: boolean;
-    arbeidserfaringer: string[];
-    totalErfaring: string[];
-}
-
-interface Typeahead {
-    value: string;
-    suggestions: string[];
-}
-
-interface TypeaheadState {
-    kompetanse: Typeahead;
-    stilling: Typeahead;
-    arbeidserfaring: Typeahead;
-    utdanning: Typeahead;
-    geografi: Typeahead;
-    geografiKomplett: Typeahead;
-    sprak: Typeahead;
-    forerkort: Typeahead;
-    navkontor: Typeahead;
-}
 
 export default AppState;
